Added DungeonStore/room generation.
